+++ gemalaya/gemqti.py
@@ -1,5 +1,9 @@
+import json
+import os
+import os.path
 import tempfile
 import traceback
+import signal
 import string
 import webbrowser
 
@@ -9,6 +13,9 @@ from omegaconf import OmegaConf
 
 import ignition
 
+from PySide6.QtCore import QDir
+from PySide6.QtCore import QFile
+from PySide6.QtCore import QIODeviceBase
 from PySide6.QtCore import QObject
 from PySide6.QtCore import QJsonValue
 from PySide6.QtCore import Slot
@@ -28,21 +35,23 @@ threadpool = concurrent.futures.ThreadPoolExecutor(max_workers=16)
 
 
 def tSlot(*args, **kws):
-    def got_result(sig, future):
+    def got_result(sig, sige, future):
         try:
             sig.emit(future.result())
-        except Exception:
-            pass
+        except Exception as err:
+            sige.emit(str(err))
 
     def outer_decorator(fn):
         @Slot(*args)
         @functools.wraps(fn)
         def wrapper(*args, **kwargs):
-            sigs = kws.get('sigSuccess')
-            sig = getattr(args[0], sigs)
+            sig = getattr(args[0], kws.get('sigSuccess'))
+
+            opt = kws.get('sigError')
+            sige = getattr(args[0], opt) if opt else None
 
             f = threadpool.submit(fn, *args)
-            f.add_done_callback(functools.partial(got_result, sig))
+            f.add_done_callback(functools.partial(got_result, sig, sige))
 
         return wrapper
     return outer_decorator
@@ -51,6 +60,9 @@ def tSlot(*args, **kws):
 class GeminiInterface(QObject):
     srvResponse = Signal(dict, arguments=['resp'])
     srvError = Signal(str, arguments=['message'])
+
+    fileDownloaded = Signal(dict, arguments=['resp'])
+    fileDownloadError = Signal(str, arguments=['error'])
 
     def __init__(self, parent=None):
         super().__init__(parent)
@@ -78,7 +90,7 @@ class GeminiInterface(QObject):
             traceback.print_exc()
             return None
 
-    @Slot(str, QJsonValue, result=str)
+    @tSlot(str, QJsonValue, sigSuccess="fileDownloaded")
     def downloadToFile(self,
                        href: str,
                        options: QJsonValue):
@@ -89,13 +101,23 @@ class GeminiInterface(QObject):
                 ca_cert=(self.certp, self.keyp)
             )
 
-            with tempfile.NamedTemporaryFile(mode='wb',
-                                             delete=False) as file:
-                file.write(response.data())
+            data = response.raw_body
 
-            return file.name
+            if response.is_a(ignition.SuccessResponse):
+                with tempfile.NamedTemporaryFile(mode='wb',
+                                                 delete=False) as file:
+                    file.write(
+                        data if isinstance(data, bytes) else data.encode())
+            else:
+                raise Exception(f'Download error for: {href}')
+
+            return {
+                'meta': response.meta,
+                'path': file.name
+            }
         except Exception:
             traceback.print_exc()
+            raise
 
     @Slot(str, str, result=str)
     def buildUrl(self,
@@ -121,6 +143,7 @@ class GeminiInterface(QObject):
                 timeout=10
             )
             assert response
+
             gemText = response.data()
 
             if response.is_a(ignition.InputResponse):
@@ -173,6 +196,17 @@ class GeminiInterface(QObject):
                 if line.type == GmiLineType.LINK:
                     # Compute keyboard sequence shortcut
 
+                    url = URL(line.extra)
+                    if url.scheme in ['http', 'https'] and \
+                            self.app.levior_proc:
+                        # Proxy web links to levior if it's running
+                        upath = url.path if url.path else '/'
+                        url = URL.build(
+                            scheme='gemini',
+                            host='localhost',
+                            path=f'/{url.host}{upath}'
+                        )
+
                     lsec, rem = divmod(linkno, len(string.ascii_letters))
 
                     if lsec == 0:
@@ -191,12 +225,12 @@ class GeminiInterface(QObject):
 
                     model.append({
                         'type': 'link',
-                        'href': line.extra,
+                        'href': str(url),
                         'hrefPrev': hrefPrev,
                         'keyseq': keyseq,
                         'title': line.text if line.text else line.extra
                     })
-                    hrefPrev = line.extra
+                    hrefPrev = str(url)
                     linkno += 1
                 elif line.type == GmiLineType.REGULAR:
                     model.append({
@@ -260,11 +294,19 @@ class GeminiInterface(QObject):
 
 
 class GemalayaInterface(QObject):
-    def __init__(self, cfg_path: Path, config, parent=None):
+    def __init__(self,
+                 cfg_dir_path: Path,
+                 cfg_path: Path,
+                 config,
+                 parent=None):
         super().__init__(parent)
 
+        self.app = QApplication.instance()
         self.config = config
+        self.cfg_dir_path = cfg_dir_path
         self.cfg_path = cfg_path
+
+        self.localt_path = self.cfg_dir_path.joinpath('themes')
 
     def __save_config(self):
         OmegaConf.save(config=self.config, f=str(self.cfg_path))
@@ -273,10 +315,97 @@ class GemalayaInterface(QObject):
     def getConfig(self):
         return OmegaConf.to_container(self.config)
 
+    @Slot(result=list)
+    def themesNames(self):
+        themes = []
+        for theme in QDir(':/gemalaya/themes').entryList():
+            themes.append(theme)
+
+        for ent in self.localt_path.glob('*'):
+            if ent.is_dir():
+                themes.append(ent.name)
+
+        return themes
+
+    @Slot(str, result=dict)
+    def getThemeConfig(self, themeName: str):
+        """
+        Return the config for a theme (custom theme from a file or qrc builtin)
+
+        :rtype: dict
+        """
+
+        theme_default_path = Path(os.path.dirname(__file__)).joinpath(
+            'theme_default.yaml'
+        )
+
+        # custom theme (local file)
+        ctp = QFile(f'{self.cfg_dir_path}/themes/{themeName}/{themeName}.yaml')
+        if ctp.exists():
+            tfile = ctp
+        else:
+            tfile = QFile(f':/gemalaya/themes/{themeName}/{themeName}.yaml')
+        try:
+            tfile.open(QIODeviceBase.ReadOnly)
+            data = tfile.readAll().data().decode()
+            theme = OmegaConf.create(data)
+
+            inherits = theme.get('inherits')
+            if inherits:
+                parent = self.getThemeConfig(inherits)
+                if not parent:
+                    raise ValueError(f'Invalid theme heritage: {inherits}')
+
+                tcfg = OmegaConf.merge(
+                    OmegaConf.create(parent),
+                    theme
+                )
+            else:
+                # merge with the defaults
+                with open(theme_default_path, 'rt') as fd:
+                    tcfg = OmegaConf.merge(
+                        OmegaConf.load(fd),
+                        theme
+                    )
+
+            if 0:
+                print(json.dumps(OmegaConf.to_container(tcfg), indent=4))
+
+            return OmegaConf.to_container(tcfg)
+        except Exception:
+            print(f'Error loading theme {themeName}: {traceback.format_exc()}')
+
+    @Slot(str, str, result=str)
+    def getThemeRscPath(self, themeName: str, path: str):
+        """
+        Returns the URL for a theme's resource
+
+        :rtype: str
+        """
+        trp = QFile(f'{self.cfg_dir_path}/themes/{themeName}/{path}')
+        if trp.exists():
+            return trp.fileName()
+
+        trp = QFile(f'qrc:/gemalaya/themes/{themeName}/{path}')
+        if trp.exists():
+            return f'qrc:/gemalaya/themes/{themeName}/{path}'
+
+        # Use sinister as a backup for now
+        return f'qrc:/gemalaya/themes/sinister/{path}'
+
     @Slot(str, QJsonValue, result=bool)
     def set(self, attr: str, value: QJsonValue):
+        """
+        Set the value for a config setting
+        """
+        cur = self.config
         try:
-            setattr(self.config, attr, value.toVariant())
+            sections = attr.split('.')
+            for s in sections[0:-1]:
+                cur = cur.get(s)
+                assert cur is not None
+
+            setattr(cur, sections[-1], value.toVariant())
         except AttributeError:
             traceback.print_exc()
             return False
@@ -289,6 +418,9 @@ class GemalayaInterface(QObject):
 
     @Slot(str)
     def browserOpenUrl(self, urlString: str):
+        """
+        Open an external url with the webbrowser API
+        """
         try:
             url = URL(urlString)
             assert url.scheme in ['http', 'https']
@@ -296,3 +428,18 @@ class GemalayaInterface(QObject):
             webbrowser.open(str(url), new=2)
         except Exception:
             traceback.print_exc()
+
+    @Slot(result=bool)
+    def httpGatewayActive(self):
+        return self.app.levior_proc is not None
+
+    @Slot()
+    def quit(self):
+        if self.app.levior_proc:
+            print(f'Stopping levior (PID: {self.app.levior_proc.pid})')
+
+            self.app.levior_proc.terminate()
+            self.app.levior_proc.send_signal(signal.SIGTERM)
+            self.app.levior_proc.kill()
+
+        self.app.quit()