+++ gemalaya/qml/GemalayaWindow.qml
@@ -44,6 +44,80 @@ Window {
     }
   }
 
+  Popup {
+    id: themeChangedPopup
+    x: window.width * 0.25
+    y: window.height * 0.5
+    margins: 20
+
+    property string message
+
+    background: Rectangle {
+      color: 'black'
+      border.color: 'white'
+      border.width: 2
+      radius: 10
+    }
+
+    contentItem: Text {
+      text: themeChangedPopup.message
+      color: 'blue'
+      font.family: 'Arial'
+      font.pointSize: 40
+    }
+  }
+
+  Scheduler {
+    id: sched
+  }
+
+  Action {
+    id: themesCycleAction
+    shortcut: Conf.shortcuts.themesCycle
+
+    onTriggered: {
+      var nextTheme
+      let ctidx = Conf.themesNames.indexOf(Conf.ui.theme)
+
+      if ((ctidx + 1) <= Conf.themesNames.length - 1)
+        nextTheme = Conf.themesNames[ctidx+1]
+      else
+        nextTheme = Conf.themesNames[0]
+
+      Conf.changeTheme(nextTheme)
+
+      themeChangedPopup.message = 'Current theme: ' + nextTheme
+      themeChangedPopup.open()
+
+      sched.delay(function() { themeChangedPopup.close() }, 1000)
+    }
+  }
+
+  Action {
+    shortcut: Conf.shortcuts.fontSizeIncrease
+    onTriggered: {
+      Conf.set('ui.fonts.defaultPointSize',
+               Conf.fontPrefs.defaultPointSize + 2)
+      Conf.set('ui.fonts.text.pointSize',
+               Conf.fontPrefs.text.pointSize + 2)
+      Conf.set('ui.fonts.links.pointSize',
+               Conf.fontPrefs.links.pointSize + 2)
+      Conf.update()
+    }
+  }
+  Action {
+    shortcut: Conf.shortcuts.fontSizeDecrease
+    onTriggered: {
+      Conf.set('ui.fonts.defaultPointSize',
+               Conf.fontPrefs.defaultPointSize - 2)
+      Conf.set('ui.fonts.text.pointSize',
+               Conf.fontPrefs.text.pointSize - 2)
+      Conf.set('ui.fonts.links.pointSize',
+               Conf.fontPrefs.links.pointSize - 2)
+      Conf.update()
+    }
+  }
+
   ColumnLayout {
     anchors.fill: parent