+++ gemalaya/qml/HeadingItem.qml
@@ -3,6 +3,7 @@ import QtQuick.Controls 2.15
 import QtQuick.Layouts 1.4
 
 Text {
+  id: heading
   property string content
   property string hsize: 'h1'
 
@@ -13,10 +14,9 @@ Text {
   KeyNavigation.priority: KeyNavigation.BeforeItem
   KeyNavigation.tab: nextLinkItem
 
-  Layout.margins: 10
+  Layout.margins: 5
   Layout.maximumWidth: width
   Layout.alignment: Qt.AlignHCenter
-  width: width
 
   signal focusRequested()
 
@@ -26,11 +26,11 @@ Text {
     font.pointSize: {
       switch(hsize) {
         case 'h1':
-          return Conf.heading.h1.fontSize
+          return Conf.fontPrefs.defaultPointSize * 1.5
         case 'h2':
-          return Conf.heading.h2.fontSize
+          return Conf.fontPrefs.defaultPointSize * 1.3
         case 'h3':
-          return Conf.heading.h3.fontSize
+          return Conf.fontPrefs.defaultPointSize * 1.1
         default:
           return 16
       }
@@ -54,7 +54,6 @@ Text {
 
   text: textmn.text
   font: textmn.font
-
   antialiasing: true
-  wrapMode: Text.WordWrap
+  wrapMode: Text.WrapAnywhere
 }