+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.3.4] - 2023-08-20
+
+### Added
+
+gemalaya:
+
+- Support for multimedia content
+- Spawn a levior process from gemalaya to be able to load web content
+  with the http to gemini proxy
+- Add shortcuts to increase/decrease the font sizes
+
 ## [0.3.1] - 2023-07-05
 
 ### Added